+++ src/EditPage/EditPage.js
@@ -13,13 +13,18 @@ export function EditPage(props){
         props.toPage(PAGE.START);
     }
 
+    const isSantaListValid = () => {
+        return props.santas.length >= 2;
+    };
+    
+
     return (
         <div className='EditPage'>
             <SantaInput addSanta={props.addSanta} />
             <SantaTable santas={props.santas} removeSanta={props.removeSanta} />
             <div className='ButtonBar'>
                 <button onClick={handleBack}>Back</button>
-                <button onClick={handleConfirm}>Confirm</button>
+                <button onClick={handleConfirm} disabled={!isSantaListValid()}>Confirm</button>
             </div>
         </div>
     );