+++ src/StartPage/StartPage.js
@@ -1,8 +1,13 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { logo } from "../constants";
 import "./StartPage.css";
 
 export function StartPage(props) {
+    // Reset santas, when the start screen is loaded
+    useEffect(() => {
+        props.removeAllSantas();
+    }, []);
+
     return (
         <div className='Start'>
             <img src={logo} alt="logo" />