+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.2.1] - 2021-12-7
+### Added
+- Validation check before enabling "Confirm" on "Edit Page"
+
+### Changed
+- Santa list will now be reset whenever the start screen is loaded
 
 ## [0.2] - 2021-12-7
 ###  Added